fix: 경험 목록 조회 시 sort_order 미반영 버그 수정

+++ src/main/java/com/kusitms/kkium/experience/controller/ExperienceController.java
@@ -62,7 +62,7 @@ public class ExperienceController {
       @AuthenticationPrincipal CustomUserDetails userDetails,
       @RequestParam(required = false) PieceType type,
       @RequestParam(required = false) String keyword,
-      @RequestParam(required = false) Long cursor,
+      @RequestParam(required = false) Integer cursor,
       @RequestParam(defaultValue = "10") @Min(1) @Max(100) int size) {
     ExperienceListResponse response =
         experienceService.getList(userDetails.getId(), type, cursor, size, keyword);
