+++ src/main/java/com/kusitms/kkium/experience/dto/response/ExperienceListResponse.java
@@ -3,4 +3,4 @@ package com.kusitms.kkium.experience.dto.response;
 import java.util.List;
 
 public record ExperienceListResponse(
-    boolean hasNext, Long nextCursor, List<ExperienceCardResponse> experiences) {}
+    boolean hasNext, Integer nextCursor, List<ExperienceCardResponse> experiences) {}
